remove class Event duplication
